fix: use f64 in common::math::factorial
This allows using much bigger numbers as the input, as well as fixes a
panic where the i64 overflows.

+++ calculator/src/astgen/ast.rs
@@ -270,7 +270,7 @@ impl AstNode {
             match m {
                 AstNodeModifier::Factorial => {
                     expect_int!(value, self.range, m);
-                    *value = math::factorial(*value as i64) as f64;
+                    *value = math::factorial(*value);
                 }
                 AstNodeModifier::BitwiseNot => {
                     expect_int!(value, self.range, m);
